Moved things around to make the project cleaner
