+++ src/components/screens/WelcomeScreen.js
@@ -3,7 +3,7 @@ import { View, Text, StatusBar, Platform } from 'react-native';
 import AppIntroSlider from 'react-native-app-intro-slider';
 import LinearGradient  from 'react-native-linear-gradient';
 import Ionicons from 'react-native-vector-icons/Ionicons';
-import { slides } from '../../../config';
+import { slides } from '../../../config/config';
 import { LoginNavigator } from '../../constants/screenNames';
 import updateNavigation from '../NavigationHelper';
 import { welcomeStyles as styles, statusBarDark } from '../../styles';
